fix: resolve code quality issues and improve error handling
- Fix type mismatch in slog logging (int64 vs int) in actions.go
- Add missing documentation for ErrorType constants in facts.go
- Refactor high-complexity function into smaller, focused helpers
- Fix unchecked os.Chdir errors in test files with proper error handling
- Remove unnecessary else blocks in SSH authentication setup
- Clean up go.mod dependencies (move uuid to indirect)
- Improve nil pointer safety in test configurations

All linting tools now pass with zero issues:
- golangci-lint: ✅ 0 issues
- staticcheck: ✅ 0 issues
- golint: ✅ 0 issues
- errcheck: ✅ 0 issues
- ineffassign: ✅ 0 issues

This commit brings the codebase to production-ready quality standards.

+++ commands/actions.go
@@ -1523,8 +1523,8 @@ func (ae *ActionExecutor) createProgressReporter(action *schemas.ActionsActionV1
 				slog.String("action", action.Description),
 				slog.String("current_file", progress.CurrentFile),
 				slog.String("operation", progress.CurrentOperation),
-				slog.Int("files_processed", progress.FilesProcessed),
-				slog.Int("total_files", progress.TotalFiles),
+				slog.Int64("files_processed", progress.FilesProcessed),
+				slog.Int64("total_files", progress.TotalFiles),
 				slog.Float64("percentage", progress.Percentage),
 				slog.Int64("bytes_transferred", progress.BytesTransferred),
 				slog.Int64("bytes_saved", progress.BytesSaved))
